Refactor/Moved class code to better name spaces
