+++ .github/workflows/test_integration.yml
@@ -66,7 +66,7 @@ jobs:
           f.write(base64.b64decode(data))
 
     - name: Extract config.zip
-      run: unzip config.zip
+      run: unzip -qq config.zip
 
     - name: Create plugin dir
       run: |
