+++ ASFFreeGames/ASFExtentions/Bot/BotContext.cs
@@ -2,9 +2,12 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using ArchiSteamFarm.Steam;
+using ASFFreeGames.ASFExtentions.Games;
+using Maxisoft.ASF;
 
-namespace Maxisoft.ASF;
+namespace ASFFreeGames.ASFExtentions.Bot;
+
+using Bot = ArchiSteamFarm.Steam.Bot;
 
 internal sealed class BotContext : IDisposable {
 	private const ulong TriesBeforeBlacklistingGameEntry = 5;
