+++ ASFFreeGames/Redlib/ParserIndices.cs
@@ -1,0 +1,3 @@
+﻿namespace Maxisoft.ASF.Redlib;
+
+internal readonly record struct ParserIndices(int StartOfCommandIndex, int EndOfCommandIndex, int StartOfFooterIndex, int HrefStartIndex, int HrefEndIndex);
